Replace old dashboard with modern shadcn/ui implementation
- Remove old dashboard directory and dashboard-sidebar component
- Rename dashboard-new to dashboard for seamless authentication flow
- Update all route references from /dashboard-new/* to /dashboard/*
- Create missing Tasks and Settings pages with placeholder content
- Fix Framer Motion import in calendar-check icon component
- Fix TypeScript type definitions in team-switcher component
- Update preload hooks to use new dashboard components
- Clean up unused imports in app-sidebar component

Benefits:
- Modern UI with shadcn/ui components and responsive sidebar
- Keyboard shortcuts (Ctrl+B / Cmd+B) for sidebar toggle
- Better mobile navigation with sheet-based drawer
- Enhanced UX with animations and consistent theming
- Cleaner, more maintainable codebase
