--- src/app/dashboard-new/layout.tsx
@@ -1,22 +1,0 @@
-import type React from "react";
-import { SidebarProvider } from "@/components/ui/sidebar";
-import { DashboardSidebar } from "@/components/app-sidebar";
-
-export default function DashboardNewLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="dark">
-      <SidebarProvider>
-        <div className="flex h-screen w-full bg-background text-foreground">
-          <DashboardSidebar />
-          <main className="flex-1 overflow-auto">
-            {children}
-          </main>
-        </div>
-      </SidebarProvider>
-    </div>
-  );
-}
