--- src/app/dashboard-new/page.tsx
@@ -1,10 +1,0 @@
-export default function DashboardNewHome() {
-  return (
-    <div className="p-6">
-      <h1 className="text-2xl font-semibold mb-4">Dashboard Home</h1>
-      <p className="text-muted-foreground">
-        Welcome to the new dashboard design using shadcn components.
-      </p>
-    </div>
-  );
-}
