+++ src/app/dashboard/page.tsx
@@ -1,13 +1,10 @@
-"use client"
-
-import { DashboardPageWrapper } from "@/components/features/dashboard/dashboard-page-wrapper"
-
-export default function Dashboard() {
+export default function DashboardNewHome() {
   return (
-    <DashboardPageWrapper>
-      <div>
-        {/* Spotlight content goes here */}
-      </div>
-    </DashboardPageWrapper>
-  )
+    <div className="p-6">
+      <h1 className="text-2xl font-semibold mb-4">Dashboard Home</h1>
+      <p className="text-muted-foreground">
+        Welcome to the new dashboard design using shadcn components.
+      </p>
+    </div>
+  );
 }
