+++ src/components/icons/calendar-check.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { motion, useAnimation } from "motion/react";
-import type { Variants } from "motion/react";
+import { motion, useAnimation } from "framer-motion";
+import type { Variants } from "framer-motion";
 import type { HTMLAttributes } from "react";
 import { forwardRef, useCallback, useImperativeHandle, useRef } from "react";
 import { cn } from "@/lib/utils";
